requirements.txt
numpy==1.26.4
opencv-python-headless==4.11.0.86
torch==1.9.0
torchvision==0.10.0
Pillow==11.1.0
PyQt5==5.15.9
pyqt5-tools==5.15.9.3.3
efficientnet-pytorch==0.7.1
retinaface-pytorch==0.0.8
tqdm==4.67.1
